More debugging in scripts, better defaults.

+++ scripts/build.sh
@@ -1,26 +1,45 @@
+source "$ROOT/rec/scripts/base.sh"
+
 CONFIG=$1
 shift
+if [ ! "$CONFIG" ]; then
+  CONFIG=debug
+fi
 
-while (( "$#" )); do
-  PACKAGE="$1"
-  shift
+if [ ! "$@" ]
+then
 
-  cd "$ROOT"
-  source "rec/scripts/variables.sh"
+  "$ROOT/rec/scripts/build.sh" $CONFIG $PACKAGES
 
-  if [ -d "$INSTALL_DIR" ] ; then
-    echo "Removing $INSTALL_DIR"
-    rm -Rf "$INSTALL_DIR"
-  fi
+else
 
-  echo "Creating $INSTALL_DIR"
-  mkdir -p "$INSTALL_DIR"
+  while (( "$#" )); do
+    PACKAGE="$1"
+    shift
+    echo
+    echo
+    echo "       *** building $PACKAGE ***"
+    echo
+    echo
 
-  echo "Building package $PACKAGE, config=$CONFIG into $INSTALL_DIR"
-  cd "$ROOT/$PACKAGE"
-  source "$SCRIPTS/configure.sh"
-  make clean
-  make
-  make install
+    cd "$ROOT"
+    source "rec/scripts/variables.sh"
 
-done
+    if [ -d "$INSTALL_DIR" ] ; then
+      echo "Removing $INSTALL_DIR"
+      rm -Rf "$INSTALL_DIR"
+    fi
+
+    echo "Creating $INSTALL_DIR"
+    mkdir -p "$INSTALL_DIR"
+
+    echo "  ===== BUILDING ===== $PACKAGE, config $CONFIG, on $PLATFORM, into $INSTALL_DIR"
+    cd "$ROOT/$PACKAGE"
+    source "$ROOT/rec/scripts/configure.sh"
+    make clean
+    make
+    make install
+
+  done
+
+fi
